backend/requirements.txt
annotated-types==0.7.0
nba_api==1.6.1
Flask==3.1.0
gunicorn==23.0.0
requests==2.32.3
python-dotenv==1.0.1
anyio==4.6.2.post1
blessed==1.20.0
blinker==1.9.0
botocore==1.35.99
cement==2.10.14
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.5
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.3
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
nba_api==1.6.1
numpy==1.26.4
packaging==24.2
panadas==0.2
pandas==2.2.3
pathspec==0.10.1
pillow==11.0.0
pydantic==2.10.1
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.32.1
wcwidth==0.2.13
Werkzeug==3.1.3
